requirements.txt
Django~=4.0.1
Pillow~=9.0.0
djangorestframework~=3.13.1
pytest~=6.2.5